+++ Assets/Mirror/Examples/LagCompensation/ServerCube.cs
@@ -1,8 +1,7 @@
 using System.Collections.Generic;
-using Mirror;
 using UnityEngine;
 
-namespace RollbackDemo
+namespace Mirror.Examples.LagCompensation
 {
     public class ServerCube : MonoBehaviour
     {
