+++ Assets/Mirror/Examples/LagCompensation/Snapshot3D.cs
@@ -1,8 +1,7 @@
 // a simple snapshot with timestamp & interpolation
-using Mirror;
 using UnityEngine;
 
-namespace RollbackDemo
+namespace Mirror.Examples.LagCompensation
 {
     public struct Snapshot3D : Snapshot
     {
